// src/interfaces/transfer.interface.ts
export interface Transfer {
  transfer_id: string;
  meeting_log_id: string;
  transfer_type: string;
  unique_member_id: string;
  ik_number: string;
  new_unique_member_id: string;
  new_ik_number: string;
  transfer_status: number;
  new_member_details_id: string;
  date_logged: string;
  staff_id: string;
  hub: string;
  app_version: string;
  imei: string;
}